дз

+++ Homework/Lesson 7/Task 2.py
@@ -1,6 +1,6 @@
-a = int(input())
-b = int(input())
-c = int(input())
+a = float(input())
+b = float(input())
+c = float(input())
 if a == b == c:
     print('Равносторонний')
 elif a == b or b == c or a == c:
